+++ greedy-string-tiling/quickGST.py
