+++ pyproject.toml
@@ -14,6 +14,8 @@ pandas = "^2.1.1"
 lizard = "^1.17.10"
 pyspellchecker = "^0.8.1"
 scikit-learn = "^1.4.1.post1"
+tqdm = "^4.66.2"
+bs4 = "^0.0.2"
 
 
 [build-system]
