+++ token-files/ArrayExamples.java.txt
